requirements.txt
streamlit
google-generativeai>=0.5.0
python-dotenv
PyPDF2
Pillow
